refactors general dbconnect config to separate package
